\name{RunHOPSApp}
\alias{RunHOPSApp}
\title{
	UK Biobank HOPS (HOrizontal Pleiotropy Score) results.
}
\description{
	Visualize and download the full set of HOPS (HOrizontal Pleiotropy Score) results on the UK Biobank.
}
\usage{
	RunHOPSApp()
}
\references{
    Daniel Jordan*, Marie Verbanck* and Ron Do. Pervasive pleiotropy driven by extreme polygenicity in the human genome revealed by a novel quantitative analysis of summary association statistics. BioRxiv.
}
\author{
    Marie Verbanck
}
\examples{
	# Visualize and download the full set of HOPS results in the UK Biobank
	RunHOPSApp()
}