\name{RunPleiotropyScoreApp}
\alias{RunPleiotropyScoreApp}
\title{
	UK Biobank pleiotropy score results.
}
\description{
	Visualize and download the full set of pleiotropy score results on the UK Biobank.
}
\usage{
	RunPleiotropyScoreApp()
}
\references{
    Daniel Jordan*, Marie Verbanck* and Ron Do. Pervasive pleiotropy driven by extreme polygenicity in the human genome revealed by a novel quantitative analysis of summary association statistics. BioRxiv.
}
\author{
    Marie Verbanck
}
\examples{
	# Visualize and download the full set of pleiotropy score results in the UK Biobank
	RunPleiotropyScoreApp()
}