\name{RunHOPSApp}
\alias{RunHOPSApp}
\title{
	UK Biobank HOPS (HOrizontal Pleiotropy Score) results.
}
\description{
	Visualize and download the full set of HOPS (HOrizontal Pleiotropy Score) results on the UK Biobank.
}
\usage{
	RunHOPSApp()
}
\references{
    Daniel Jordan*, Marie Verbanck* and Ron Do. HOPS: a quantitative score reveals pervasive horizontal pleiotropy in human genetic variation is driven by extreme polygenicity of human traits and diseases. Genome Biology 20, 222 (2019).
}
\author{
    Marie Verbanck
}
\examples{
	# Visualize and download the full set of HOPS results in the UK Biobank
	RunHOPSApp()
}