\name{RunPleiotropyScoreApp}
\alias{RunPleiotropyScoreApp}
\title{
	UK Biobank pleiotropy score results.
}
\description{
	Visualize and download the full set of pleiotropy score results on the UK Biobank.
}
\usage{
	RunPleiotropyScoreApp()
}
\references{
    Daniel Jordan*, Marie Verbanck* and Ron Do. HOPS: a quantitative score reveals pervasive horizontal pleiotropy in human genetic variation is driven by extreme polygenicity of human traits and diseases. Genome Biology 20, 222 (2019).
}
\author{
    Marie Verbanck
}
\examples{
	# Visualize and download the full set of pleiotropy score results in the UK Biobank
	RunPleiotropyScoreApp()
}